:hammer: Minor changes in file names. Update books JSON
